requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
py4j==0.10.9.7
    # via pyspark
pyspark==3.5.1
python-dotenv==1.0.1
requests==2.32.2
urllib3==2.2.1
    # via requests
